Multiple keyword trigger issue fixed.
Fixed an issue with keywords triggering both gif and png, if there were
similiarities in the name. Is not an issue with the same file type.

Also added a stopper on both add_image and edit_image to stop any cases
where someone tries to add or edit an image that already has an existing
keyword.