Horoscope - Fix date text
On horo and chinese horo commands, the date regex was returning a lot of html prepended on the date, causing the response to be over 4000 chars. I added a separate date regex and simplified the original regex for finding the horoscope. Granted I'm not a regex pro by any means so feel free to adjust this PR if you accept it. Seemed to work fine in all my testing of various astral signs and chinese signs.